refactor: Implement complete MVC architecture
- Reorganize the project directory into models, views, and controllers
- Extract database, schema, and payroll logic into src/models/
- Separate pure GUI layouts and custom widgets into src/views/
- Implement MainController in src/controllers/ to handle business logic, database syncing, and signal routing
- Update src/main.py to serve as a lightweight application bootstrapper
- Fix internal import paths across all modules and test files
